// src/Components/types.ts
// types.ts
export interface ItemPNCP {
  id: string;
  title: string;
  description: string;
  item_url: string;
  document_type: string;
  createdAt: string;
  ano: string;
  numero_sequencial: string;
  numero_controle_pncp: string;
  orgao_nome: string;
  unidade_nome: string;
  esfera_nome: string;
  poder_nome: string;
  municipio_nome: string;
  uf: string;
  modalidade_licitacao_nome: string;
  situacao_nome: string;
  data_publicacao_pncp: string;
  data_inicio_vigencia: string;
  data_fim_vigencia: string;
  tipo_nome: string;
  orgao_cnpj: string;
}